Renamed blink, blink2, count, and fib (all .hex) to .phx (plain hex)
This renaming was needed because the ".hex" extension is widely
associated with the "Intel Hex" format. Since the new version of
Run_1802 will also support Intel Hex format, it makes sense to adopt
the widely used extension for that format (.hex) rather than try to
retain the ".hex" format for "plain hex".