Run_1802.py and Wiring Diagram with Data GPIO (0-7) Addr GPIO (8-15)
This verion of Run_1802.py appears to work properly with the data and
address mapping presented on February 27, 2023 in the COSMACELF post
number #40363. This mapping assigns the 1802's data bus pins (0-7) to
the Raspberry Pi's GPIO pins (0-7). It also assigns the 1802's address
pins (0-7) to the Raspberry Pi's GPIO pins (8-15). Consult the code
for pin assignments. This version also adds a "c=#.##" parameter which
sets the clock timing to the floating point number equated to "c".