Updated Run_1802 to provide a minimal instruction trace
This version borrowed from some earlier Javascript code to produce
a minimal trace of the instructions fetched during a run. Each
instruction is displayed in hex along with its mnemonic string.
Registers are shown for instructions with an implicit register
reference, but immediate data is not shown. The trace looks like:

  Fetch at addr 0008 got 1e = INC E
  Fetch at addr 0009 got 8e = GLO E
  Fetch at addr 000a got 5f = STR F
  Fetch at addr 000b got 64 = OUT4
  7
  Fetch at addr 000c got 2f = DEC F
  Fetch at addr 000d got 30 = BR
  Fetch at addr 0008 got 1e = INC E
  Fetch at addr 0009 got 8e = GLO E
  Fetch at addr 000a got 5f = STR F
  Fetch at addr 000b got 64 = OUT4
  8
  Fetch at addr 000c got 2f = DEC F
  Fetch at addr 000d got 30 = BR
  Fetch at addr 0008 got 1e = INC E

Note that output data (port 4) is interleaved with the trace
data. In this example, the program is counting, so the 7 and 8
are from the "OUT4" instructions just prior to their display.