Separated Pin Definitions into Pi_to_1802.py and added operand display
This version separates the Pin Definitions into a separate source file
named "Pi_to_1802.py". This makes it easier to adapt the code to other
pin definitions.

This version also displays operands as part of the trace output. These
operands are actually from the virtual memory image since they have not
yet been sent to the 1802 during the fetch cycle when the trace data is
printed. This produces the output shown here:

  $ python Run_1802.py n=500 f=test_addr_long.ahx t

  Loading:
  ========================================
  ; Blink, Jump and repeat

  00:     ; Locate this code at 00
  7a      ; REQ
  c0 fffc ; LBR fffc

  fffc:   ; Locate this code at fffc
  7b      ; SEQ
  c0 0000 ; LBR 00
  ========================================

  Running 500 clocks
  Fetch at addr 0000 got 7a = REQ
  Fetch at addr 0001 got c0 = LBR ff fc
  Fetch at addr fffc got 7b = SEQ
  Fetch at addr fffd got c0 = LBR 00 00
  Fetch at addr 0000 got 7a = REQ
  Fetch at addr 0001 got c0 = LBR ff fc
  Fetch at addr fffc got 7b = SEQ
  Fetch at addr fffd got c0 = LBR 00 00
  Fetch at addr 0000 got 7a = REQ
  Fetch at addr 0001 got c0 = LBR ff fc
  Fetch at addr fffc got 7b = SEQ
  Fetch at addr fffd got c0 = LBR 00 00

Providing the operands (branch addresses in this case) makes
the trace much easier to follow.