Added code to create a known state before recording of reset data.
Previous versions recorded data prior to reset which depended on the
state of the 1802 processor before reset was asserted. These most
often appeared as the previous values of the address bus prior to
resetting. These lingering address values were reset to zero during
the 9 clock cycles of the reset process, but they showed up as random
values in the output log for the run. This would cause differences to
show when comparing with other runs which would otherwise be identical.

This problem was solved by implementing two resets. The first reset
is followed by a few clock cycles (currently 14) which begins to
execute the code loaded at memory address 0. This is enough time to
cause the address lines to be set to 0. At that point, a second reset
is issued which will begin the reset process again with the address
lines already set to 0. This is the data that will be captured in the
log for comparison with other runs.