Added a number of images for reference and historical perspective.
The community approved printed circuit board for the 1802 in PiLand
project ended up crossing Q and /CLEAR lines by mistake. Since this
PCB was manufactured and distributed to a number of members of the
group, it became the standard. The file:

  Agreed_1802_to_GPIO_2023_03_03c_Tony_Board.png

documents this change. Additionally, the file:

  Protection_Resistor_Mapping.png

documents the signals present on each of the protection resistors
included on the board. These are convenient points for testing.