Added GUI Code (previous commit) and support for "NoPi" mode
Note: The previous commit included addition of the initial version
of the Run_1802 GUI using TKinter. This wasn't documented in the
commit message.

This modification adds support for the "FakeGPIO" module via NoPi.

The Run_1802.py program normally runs on a Raspberry Pi which has
access to the RPi.GPIO module. However, useful development of the
GUI can be done without an actual GPIO module. This change allows
Run_1802.py to run on platforms that do not include GPIO (such as
a typical Linux workstation or PC). The "NoPi" command line
argument has been added to explicitly request this mode of
operation.