Added enough 1802 simulation to provide N2 (I/O port 4) output of 1802
This version simulates the 1802 outputting the values 1 8 0 2 in a loop.
The values are output with N2 high (I/O port 4).
Note that these are produced without any actual instructions requesting
output. They are also produced on a fixed number of clock pulses without
regard to whether that is representative of actual 1802 behavior.