Modified to use the values in RAM as "fake" output from a fake 1802.
When running with NoPi, there is no 1802, so a "fake" 1802 is used.
However, this fake 1802 doesn't need to do anything but produce the
output needed by Run_1802.py. An easy way to get this output is from
what would be the program that's normally loaded into RAM. So rather
than load a program, the fake 1802 can simply load whatever data is
in RAM and use that as the fake output from the fake 1802. This can
be used to test the various graphics_xxx.py programs without having
an actual 1802 or even a Raspberry Pi. In other words, this allows
testing on a regular PC.