Major rewrite of QixColor.asm with EQU, IF, and color options
This version of QixColor works properly in monochrome, but still
does not work in color. It was the result of a large parallel
effort to rewrite it to make better use of the available assembler
features (like EQU, IF, ELSE, ENDI). It currently has three color
modes: monochrome, all green, and full color. It's not clear why
the last two do not work, but the fact that it at least works in
monochrome makes it a suitable replacement for the earlier version.